requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.6.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
distro==1.9.0
Flask==2.3.3
Flask-WTF==1.2.1
groq==0.26.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
Markdown==3.8
MarkupSafe==3.0.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
python-docx==1.1.2
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
WTForms==3.2.1
